+++ changelog.md
@@ -3,3 +3,8 @@
 ## 1.0 (2020-05-16)
 
 First release
+
+## 1.1 (2020-05-16)
+
+* Changed namespace structure
+* Added trait `Limanweb\Models\Concerns\UsesUuid`